Refactor: 잘못된 값 입력시 좀 더 자세한 오류 메시지 출력 errors.properties를 작성했고 글로벌 에러처리도 추가 Related to: #18
