+++ src/main/resources/errors.properties
@@ -1,0 +1,17 @@
+typeMismatch.member.age = \uD68C\uC6D0 \uB098\uC774\uB294 \uC22B\uC790\uB9CC \uC785\uB825 \uAC00\uB2A5\uD569\uB2C8\uB2E4.
+typeMismatch.age = \uB098\uC774\uB294 \uC22B\uC790\uB9CC \uC785\uB825 \uAC00\uB2A5\uD569\uB2C8\uB2E4.
+typeMismatch.java.lang.Integer = \uC22B\uC790\uB9CC \uC785\uB825 \uAC00\uB2A5\uD569\uB2C8\uB2E4.
+typeMismatch = \uD0C0\uC785 \uC624\uB958 \uC785\uB2C8\uB2E4.
+
+Min.member.age = \uD68C\uC6D0 \uAC00\uC785\uC740 {1}\uC138 \uC774\uC0C1\uBD80\uD130 \uAC00\uB2A5\uD569\uB2C8\uB2E4.
+Min.age = {1}\uC138 \uC774\uC0C1\uB9CC \uAC00\uB2A5\uD569\uB2C8\uB2E4.
+Min.java.lang.Integer = {1} \uC774\uC0C1 \uC22B\uC790\uB9CC \uAC00\uB2A5\uD569\uB2C8\uB2E4.
+Min = {1} \uC774\uC0C1\uB9CC \uAC00\uB2A5\uD569\uB2C8\uB2E4.
+
+NotBlank.boardTypeCode = \uBB38\uC758\uC720\uD615\uC744 \uC120\uD0DD\uD574\uC8FC\uC138\uC694.
+NotBlank = \uD544\uC218 \uC785\uB825 \uC0AC\uD56D \uC785\uB2C8\uB2E4.
+
+NotNull.genderType = \uC131\uBCC4\uC744 \uC120\uD0DD\uD574\uC8FC\uC138\uC694.
+NotNull = \uD544\uC218 \uC785\uB825 \uC0AC\uD56D \uC785\uB2C8\uB2E4.
+
+totalLenMin = \uC81C\uBAA9\uACFC \uB0B4\uC6A9\uC758 \uCD5C\uC18C \uAE00\uC790\uC218\uB294 {0} \uC774\uC0C1 \uC774\uC5B4\uC57C \uD569\uB2C8\uB2E4. \uD604\uC7AC \uAE00\uC790\uC218 : {1}
